tests/wpt/web-platform-tests/css/css-multicol/parsing/column-fill-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: getComputedValue().columnFill</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-fill">
<meta name="assert" content="column-fill computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("column-fill", "auto");
test_computed_value("column-fill", "balance");
test_computed_value("column-fill", "balance-all");
</script>
</body>
</html>
