src/third_party/blink/web_tests/external/wpt/css/css-multicol/parsing/column-fill-computed-expected.txt

This is a testharness.js-based test.
PASS Property column-fill value 'auto' computes to 'auto'
PASS Property column-fill value 'balance' computes to 'balance'
FAIL Property column-fill value 'balance-all' computes to 'balance-all' assert_equals: expected "balance-all" but got "balance"
Harness: the test ran to completion.
